requirements.txt
aiogram>=3.0,<4.0
yandex-music>=2.0
aiohttp>=3.9
python-dotenv>=1.0
